Convert prices to using MySQL instead of sqlite.
This way we can get around our current problem of the sqlite db refusing to
write further records.

It also opens up the possibility of cross-db queries should we want prices on
the website.
