+++ SETUP
@@ -14,6 +14,9 @@
   -- CREATE USER '<mysql_user>'@'<mysql_host>' IDENTIFIED BY '<mysql_password>';
   -- GRANT ALL ON <decksite_database>.* TO '<mysql_user>'@'<mysql_host>';
   -- Use can test this worked from the commandline with `mysql <decksite_database> -u <mysql_user> -p<mysql_password>
+  -- CREATE DATABASE <prices_database>;
+  -- GRANT ALL ON <prices_database>.* TO '<mysql_user>'@'<mysql_host>';
+  -- Use can test this worked from the commandline with `mysql <prices_database> -u <mysql_user> -p<mysql_password>
 - The sqlite extension we use for fuzzy matching is not part of any string distribution so you must download and build it.
   - curl http://sqlite.org/cgi/src/raw/ext/misc/spellfix.c?name=a4723b6aff748a417b5091b68a46443265c40f0d -o spellfix.c
   - curl http://sqlite.org/cgi/src/raw/src/sqlite3ext.h?name=8648034aa702469afb553231677306cc6492a1ae -o sqlite3ext.h
