+++ shared/configuration.py
@@ -11,8 +11,7 @@ DEFAULTS = {
     'decksite_database': 'decksite',
     'image_dir': './images',
     'magic_database': './cards.sqlite',
-    'prices_database': './prices.db',
-    'pricesdb': './prices.db',
+    'prices_database': 'prices',
     'scratch_dir': '.',
     'spellfix': './spellfix',
     'to_password': '',
